changes added based in Mirek's comment
